+++ openeo/extra/spectral_indices/__init__.py
@@ -1,0 +1,13 @@
+"""
+Easily calculate spectral indices (vegetation, water, urban etc.).
+
+Supports the indices defined in the
+`Awesome Spectral Indices <https://awesome-ee-spectral-indices.readthedocs.io/>`_ project
+by `David Montero Loaiza <https://github.com/davemlz>`_.
+
+
+.. versionadded:: 0.9.1
+
+"""
+
+from openeo.extra.spectral_indices.spectral_indices import *